feat:eventList 페이징 처리 완료

+++ src/main/java/com/pickupluck/ecogging/domain/file/repository/FileRepository.java
@@ -3,5 +3,7 @@ package com.pickupluck.ecogging.domain.file.repository;
 import com.pickupluck.ecogging.domain.file.entity.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<File, Integer> {
+public interface FileRepository extends JpaRepository<File, Long> {
+
+
 }
